+++ src/main/java/br/com/fogliato/rinhabackend/repository/TransactionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TransactionRepository extends CrudRepository<TransactionEntity, Long> {
+public interface TransactionRepository extends CrudRepository<TransactionEntity, Integer> {
 
     @Query("SELECT t FROM TransactionEntity t WHERE t.customerId = ?1 ORDER BY t.createdAt Desc LIMIT 10")
     List<TransactionEntity> findLastTransactions(long customerId);
